requirements.txt
pydrive
discord.py
python-dotenv
